moving b2b recipe to setup lib
