+++ core-libs/setup/public_api.ts
@@ -1,1 +1,2 @@
-export const dummy = {};
+export * from './recipes/index';
+
