+++ core-libs/setup/recipes/b2b/b2b-storefront.module.ts
@@ -1,26 +1,24 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
+import { CostCenterModule, provideConfig, provideDefaultConfig, provideDefaultConfigFactory } from '@spartacus/core';
 import {
-  CostCenterModule,
-  provideConfig,
-  provideDefaultConfig,
-  provideDefaultConfigFactory,
-} from '@spartacus/core';
-import { CmsLibModule } from '../cms-components/cms-lib.module';
-import { StorefrontConfig } from '../storefront-config';
-import {
-  defaultB2bCheckoutConfig,
-  defaultB2bOccConfig,
+  CmsLibModule,
+  defaultCmsContentConfig,
   layoutConfig,
   mediaConfig,
-} from './config/index';
-import { defaultCmsContentConfig } from './config/static-cms-structure/default-cms-content.config';
-import { StorefrontModule } from './storefront.module';
+  StorefrontConfig,
+  StorefrontModule
+} from '@spartacus/storefront';
+import { OrganizationModule } from '@spartacus/my-account/organization';
+import { defaultB2bOccConfig } from './config/b2b-occ-config';
+import { organizationLayoutConfig } from './config/organization-layout.config';
+import { defaultB2bCheckoutConfig } from './config/b2b-checkout-config';
+
 
 @NgModule({
   imports: [
     CostCenterModule.forRoot(),
+    OrganizationModule,
     StorefrontModule,
-
     // the cms lib module contains all components that added in the bundle
     CmsLibModule,
   ],
@@ -32,10 +30,17 @@ import { StorefrontModule } from './storefront.module';
       },
     }),
     provideDefaultConfig(layoutConfig),
+    provideDefaultConfig(organizationLayoutConfig),
     provideDefaultConfig(mediaConfig),
     provideDefaultConfig(defaultB2bOccConfig),
-    provideDefaultConfig(defaultB2bCheckoutConfig),
     provideDefaultConfigFactory(defaultCmsContentConfig),
+    provideDefaultConfig(defaultB2bCheckoutConfig),
+    provideDefaultConfig({
+      pwa: {
+        enabled: true,
+        addToHomeScreen: true,
+      },
+    }),
   ],
   exports: [StorefrontModule],
 })
