+++ core-libs/setup/recipes/b2b/config/b2b-checkout-config.ts
@@ -1,8 +1,4 @@
-import {
-  CheckoutConfig,
-  DeliveryModePreferences,
-} from '../../cms-components/checkout/config/checkout-config';
-import { CheckoutStepType } from '../../cms-components/checkout/model/checkout-step.model';
+import { CheckoutConfig, CheckoutStepType, DeliveryModePreferences } from '@spartacus/storefront';
 
 export const defaultB2bCheckoutConfig: CheckoutConfig = {
   checkout: {
