+++ core-libs/setup/recipes/b2b/config/index.ts
@@ -1,2 +1,3 @@
 export * from './b2b-occ-config';
 export * from './organization-layout.config';
+export * from './b2b-checkout-config';
