+++ core-libs/setup/recipes/b2b/config/organization-layout.config.ts
