+++ core-libs/setup/recipes/b2b/index.ts
@@ -1,0 +1,2 @@
+export * from './config/index';
+export * from './b2b-storefront.module';
