+++ core-libs/setup/recipes/index.ts
@@ -1,0 +1,1 @@
+export * from './b2b/index';
