+++ feature-libs/my-account/organization/index.ts
@@ -1,2 +1,1 @@
 export * from './organization.module';
-export * from './recipes/index';
