--- feature-libs/my-account/organization/recipes/index.ts
@@ -1,2 +1,0 @@
-export * from './b2b-storefront.module';
-export * from './config/index';
