+++ projects/storefrontapp/src/environments/b2b/b2b.feature.ts
@@ -1,4 +1,4 @@
-import { B2bStorefrontModule } from '@spartacus/my-account';
+import { B2bStorefrontModule } from '@spartacus/setup';
 import { OrganizationModule } from '@spartacus/my-account/organization';
 import {
   organizationTranslationChunksConfig,
