+++ projects/storefrontlib/src/recipes/config/index.ts
@@ -1,5 +1,3 @@
-export * from './b2b-checkout-config';
-export * from './b2b-occ-config';
 export * from './default-media.config';
 export * from './layout-config';
 export * from './static-cms-structure/index';
