+++ projects/storefrontlib/src/recipes/index.ts
@@ -1,4 +1,3 @@
 export * from './b2c-storefront.module';
-export * from './b2b-storefront.module';
 export * from './config/index';
 export * from './storefront-foundation.module';
